ref(operator): restructure modules
